+++ src/main/groovy/com/ullink/NuGetSpec.groovy
@@ -1,10 +1,7 @@
 package com.ullink
 
-import groovy.util.XmlSlurper
 import groovy.util.slurpersupport.GPathResult
 import groovy.xml.XmlUtil
-import groovy.xml.MarkupBuilder
-import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.Exec
 
 class NuGetSpec extends Exec {
@@ -27,8 +24,7 @@ class NuGetSpec extends Exec {
     void generateNuspecFile() {
         def nuspecXml = generateNuspec()
         if (nuspecXml) {
-            def file = getNuspecFile()
-            file.write nuspecXml
+            getNuspecFile().write(nuspecXml, 'utf-8')
         }
     }
 
@@ -65,24 +61,71 @@ class NuGetSpec extends Exec {
     }
 
     String supplementDefaultValueOnNuspec(String nuspecString) {
+        def final msbuildTaskExists = project.tasks.findByName('msbuild') != null
+        def final packageConfigFileName = 'packages.config'
 
-        def root = new XmlSlurper(false, false).parseText(nuspecString)
+        GPathResult root = new XmlSlurper(false, false).parseText(nuspecString)
 
-        def defaultValues = {}
-        def applyDefaultValue = { String node, String value ->
+        def defaultMetadata = []
+        def setDefaultMetadata = { String node, value ->
             if (root.metadata[node].isEmpty()) {
-                defaultValues <<= { delegate."$node" value }
+                defaultMetadata.add( { delegate."$node" value } )
             }
         }
-        applyDefaultValue ('id', project.name)
-        applyDefaultValue ('version', project.version)
-        applyDefaultValue ('description', project.description)
 
-        if (root.metadata.isEmpty()) {
-            root.appendNode { metadata defaultValues }
-        } else {
-            root.metadata.appendNode defaultValues
+        setDefaultMetadata ('id', project.name)
+        setDefaultMetadata ('version', project.version)
+        setDefaultMetadata ('description', project.description ? project.description : project.name)
+
+        def appendAndCreateParentIfNeeded = {
+            String parentNodeName, List children ->
+                if(!children.isEmpty()) {
+                    if (root."$parentNodeName".isEmpty()) {
+                        root << { "$parentNodeName" children }
+                    } else {
+                        root."$parentNodeName" << children
+                    }
+                }
         }
+
+        if (msbuildTaskExists) {
+            project.logger.debug("Msbuild plugin detected. Will add defaults from it.")
+            def mainProject = project.msbuild.mainProject
+
+            if(root.files.file.isEmpty()) {
+                project.logger.debug("No files already defined in the NuGet spec, will add the ones from the msbuild task.")
+                def defaultFiles = []
+                project.msbuild.mainProject.dotnetArtifacts.each {
+                    artifact ->
+                        def fwkFolderVersion = mainProject.properties.TargetFrameworkVersion.toString().replace('v', '').replace('.', '')
+                        defaultFiles.add({ file(src: artifact.toString(), target: 'lib/net' + fwkFolderVersion) })
+                }
+                appendAndCreateParentIfNeeded('files', defaultFiles)
+            }
+
+            def packageConfigFile = new File(
+                    new File(mainProject.projectFile).parentFile,
+                    packageConfigFileName)
+            if (packageConfigFile.exists()) {
+                project.logger.debug("Adding dependencies from packages.config")
+                def defaultDependencies = []
+                def packages = new XmlParser().parse(packageConfigFile)
+                packages.package
+                        .findAll { !it.@developmentDependency.toString().toBoolean() }
+                        .each {
+                    packageElement ->
+                        defaultDependencies.add({
+                            dependency(id: packageElement.@id, version: packageElement.@version)
+                        })
+                }
+                setDefaultMetadata('dependencies', defaultDependencies)
+            }
+        }
+
+        appendAndCreateParentIfNeeded('metadata', defaultMetadata)
+
+        project.logger.info("Generated NuGetSpec file with ${root.files.file.size()} files " +
+                "and ${root.dependencies.dependecy.size()} dependencies")
         XmlUtil.serialize (root)
     }
 }